requirements.txt
Flask
google-genai
gunicorn 
Pillow
jinja2